requirements.txt
flask
flask-cors
openai
selenium
pandas
a2a-sdk
google-adk
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0